User/ma/warning level4 (#47)
* Foundation compiles with /W4

* Core compiles with /W4

* System compiles with /W4

* RendererFoundation & RendererDX11 compile with /W4

* Texture & Utilities compile with /W4

* Fixes after merge

* Renamed shader BOOL to BOOLEAN as any include of RenderContext.h before a windows.h include would conflict with the macro definition there

* Enabled warning level 4 for all tested projects so far

+++ Code/Engine/Texture/Image/Formats/WicFileFormat.cpp
@@ -145,7 +145,7 @@ ezResult ezWicFileFormat::ReadImage(ezStreamReader& stream, ezImage& image, ezLo
             const uint8_t* sourcePixels = sourceImage->pixels;
             for (ezUInt32 rowIdx = 0; rowIdx < imageHeader.GetHeight(); ++rowIdx)
             {
-              ezMemoryUtils::Copy(destPixels, sourcePixels, bytesPerRow);
+              ezMemoryUtils::Copy(destPixels, sourcePixels, static_cast<size_t>(bytesPerRow));
 
               destPixels += destRowPitch;
               sourcePixels += sourceImage->rowPitch;
